requirements.txt
requests==2.11.1
lxml==3.5.0
python-telegram-bot==10.1.0
peewee==2.6.4